requirements.txt
# Core dependencies
aiohttp>=3.11.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# UI and automation
pyautogui>=0.9.54
mss>=9.0.1
Pillow>=10.0.0
pytesseract>=0.3.10

# Jan.ai integration
requests>=2.31.0
websockets>=11.0.3

# Utilities
rich>=14.0.0
typer>=0.15.0
python-json-logger>=2.0.7

# Type checking
mypy>=1.5.1
types-requests>=2.31.0.2
types-PyYAML>=6.0.12.12

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0

numpy>=2.2.0